Added the 8 new filters
Added the Blue, Red and Green channel filters, which show how much red, green or blue is in each image, also created tint filters, which basically add a red, green or blue filter on the images. Finally added both a warhol filter and a flipped warhol filter, which pretty much creates four versions of the same picture on the panel and it applies different tints to each one. The flipped warhol does what the regular warhol does but it also flips the images direction depending on its position.